+++ src/components/LoadMoreComments.js
@@ -1,11 +1,13 @@
 const LoadMoreComments = ({
-	id,
+	data,
 	getComments,
 	perma_link,
 	setCurrentComments
 }) => {
 	const load = () => {
-		let url = "https://www.reddit.com" + perma_link + "/" + id + ".json";
+		console.log(data.id);
+		return null;
+		let url = "https://www.reddit.com" + perma_link + "/" + data.id + ".json";
 		getComments(url).then((comObj) => {
 			if (comObj === 1) {
 				alert("likely fetch request went wrong");
@@ -18,7 +20,8 @@ const LoadMoreComments = ({
 	return (
 		// todo: show how many to load
 		<button class="comments-loader" onClick={load}>
-			load more comments {id}
+			{/*load more comments */}{data.id}{" "}
+			{data.name}
 		</button>
 	);
 };