+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## CHANGELOG
 
-1. removed border-bottom and thicker border-left on comment threads.
+1. changed select subreddit shortcut to "/".
+2. added "default" to list of badThumbnails.
+3. on clicking the sub banner you go the the sub's listings.